coreFunc/bookverse/keeper/msg_server_market_test.go
package keeper

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/require"

	"github.com/goldipatel12/marketplace/x/Bookverses/types"
)

func TestMarketMsgServerCreate(t *testing.T) {
	srv, ctx := setupMsgServer(t)
	creator := "A"
	for i := 0; i < 5; i++ {
		resp, err := srv.CreateMarket(ctx, &types.MsgCreateMarket{Creator: creator})
		require.NoError(t, err)
		assert.Equal(t, i, int(resp.Id))
	}
}

func TestMarketMsgServerUpdate(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgUpdateMarket
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgUpdateMarket{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateMarket{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateMarket{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		tc := tc
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)
			_, err := srv.CreateMarket(ctx, &types.MsgCreateMarket{Creator: creator})
			require.NoError(t, err)

			_, err = srv.UpdateMarket(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}

func TestMarketMsgServerDelete(t *testing.T) {
	creator := "A"

	for _, tc := range []struct {
		desc    string
		request *types.MsgDeleteMarket
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgDeleteMarket{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgDeleteMarket{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "KeyNotFound",
			request: &types.MsgDeleteMarket{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	} {
		tc := tc
		t.Run(tc.desc, func(t *testing.T) {
			srv, ctx := setupMsgServer(t)

			_, err := srv.CreateMarket(ctx, &types.MsgCreateMarket{Creator: creator})
			require.NoError(t, err)
			_, err = srv.DeleteMarket(ctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}
